feat: Redesign public-facing UI/UX
- Redesigned the splash screen, landing page, pricing page, sign-up page, and log-in page to be more modern, visually appealing, and conversion-focused.
- Refactored all pages into reusable React components for better maintainability.
- Created a consistent and maintainable design system with global styles and design tokens.
- Added micro-interactions and animations to enhance the user experience.
- Fixed a bug in the logger that was preventing the database tests from running.
