+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test:database": "tsx scripts/test-database.ts"
+    "test:database": "tsx --require dotenv/config scripts/test-database.ts"
   },
   "dependencies": {
     "@emailjs/browser": "^4.4.1",
@@ -41,7 +41,7 @@
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
     "cross-env": "^7.0.3",
-    "dotenv": "^16.3.1",
+    "dotenv": "^16.6.1",
     "eslint": "^9",
     "eslint-config-next": "15.4.4",
     "husky": "^8.0.3",
@@ -50,7 +50,7 @@
     "lint-staged": "^15.1.0",
     "prettier": "^3.1.0",
     "tailwindcss": "^4",
-    "tsx": "^4.7.0",
+    "tsx": "^4.20.3",
     "typescript": "^5"
   },
   "optionalDependencies": {
