+++ src/app/login/page.tsx
@@ -4,23 +4,18 @@ import React, { useState } from 'react'
 import { motion } from 'framer-motion'
 import { 
   Building2, 
-  ArrowRight, 
   Mail, 
   Lock,
-  AlertCircle,
-  CheckCircle2
 } from 'lucide-react'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { Button } from '@/components/ui/Button'
 import { supabase } from '@/lib/supabaseClient'
-import { fadeInUp } from '@/lib/animations'
 
 export default function LoginPage() {
   const router = useRouter()
   const [loading, setLoading] = useState(false)
   const [error, setError] = useState('')
-  const [success, setSuccess] = useState('')
   const [formData, setFormData] = useState({
     email: '',
     password: ''
@@ -30,23 +25,16 @@ export default function LoginPage() {
     e.preventDefault()
     setLoading(true)
     setError('')
-    setSuccess('')
 
     try {
-      const { data, error: authError } = await supabase.auth.signInWithPassword({
+      const { error: authError } = await supabase.auth.signInWithPassword({
         email: formData.email,
         password: formData.password
       })
 
       if (authError) throw authError
 
-      // Show success message briefly
-      setSuccess('Login successful! Redirecting...')
-      
-      // Redirect to dashboard
-      setTimeout(() => {
-        router.push('/dashboard')
-      }, 1000)
+      router.push('/dashboard')
     } catch (err: any) {
       setError(err.message || 'Invalid email or password')
     } finally {
@@ -55,36 +43,26 @@ export default function LoginPage() {
   }
 
   return (
-    <div className="min-h-screen bg-gradient-to-br from-blue-50 via-white to-indigo-50 flex items-center justify-center px-4 sm:px-6 lg:px-8">
-      <motion.div
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ duration: 0.5 }}
-        className="w-full max-w-md"
-      >
-        {/* Logo */}
+    <div className="min-h-screen flex items-center justify-center bg-gray-50 px-4 sm:px-6 lg:px-8">
+      <div className="w-full max-w-sm">
         <Link href="/" className="flex items-center justify-center gap-2 mb-8">
-          <div className="bg-gradient-to-r from-blue-600 to-indigo-600 p-2 rounded-lg">
+          <div className="bg-blue-600 p-2 rounded-lg">
             <Building2 className="w-8 h-8 text-white" />
           </div>
           <span className="text-2xl font-bold text-gray-900">BuildLedger</span>
         </Link>
 
-        {/* Card */}
-        <div className="bg-white rounded-2xl shadow-xl p-8">
-          {/* Heading */}
-          <div className="mb-8 text-center">
-            <h1 className="text-3xl font-bold text-gray-900 mb-2">
-              Welcome back
+        <div className="bg-white rounded-2xl shadow-lg p-8">
+          <div className="mb-6 text-center">
+            <h1 className="text-2xl font-bold text-gray-900 mb-1">
+              Welcome Back
             </h1>
             <p className="text-gray-600">
-              Sign in to your account to continue
+              Log in to access your dashboard.
             </p>
           </div>
 
-          {/* Form */}
-          <form onSubmit={handleSubmit} className="space-y-6">
-            {/* Email field */}
+          <form onSubmit={handleSubmit} className="space-y-4">
             <div>
               <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">
                 Email Address
@@ -97,20 +75,19 @@ export default function LoginPage() {
                   required
                   value={formData.email}
                   onChange={(e) => setFormData({ ...formData, email: e.target.value })}
-                  className="w-full pl-10 pr-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
-                  placeholder="john@construction.com"
+                  className="w-full pl-10 pr-3 py-2 border border-gray-300 rounded-lg focus:ring-1 focus:ring-blue-500 focus:border-blue-500 transition"
+                  placeholder="you@company.com"
                 />
               </div>
             </div>
 
-            {/* Password field */}
             <div>
               <div className="flex items-center justify-between mb-1">
                 <label htmlFor="password" className="block text-sm font-medium text-gray-700">
                   Password
                 </label>
                 <Link href="/reset-password" className="text-sm text-blue-600 hover:underline">
-                  Forgot password?
+                  Forgot?
                 </Link>
               </div>
               <div className="relative">
@@ -121,85 +98,36 @@ export default function LoginPage() {
                   required
                   value={formData.password}
                   onChange={(e) => setFormData({ ...formData, password: e.target.value })}
-                  className="w-full pl-10 pr-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent transition-all"
+                  className="w-full pl-10 pr-3 py-2 border border-gray-300 rounded-lg focus:ring-1 focus:ring-blue-500 focus:border-blue-500 transition"
                   placeholder="••••••••"
                 />
               </div>
             </div>
 
-            {/* Error message */}
             {error && (
-              <motion.div
-                initial={{ opacity: 0, y: -10 }}
-                animate={{ opacity: 1, y: 0 }}
-                className="flex items-start gap-2 p-3 bg-red-50 border border-red-200 rounded-lg text-red-700 text-sm"
-              >
-                <AlertCircle className="w-5 h-5 flex-shrink-0 mt-0.5" />
-                <span>{error}</span>
-              </motion.div>
+              <div className="p-3 bg-red-50 border border-red-200 rounded-lg text-red-700 text-sm">
+                {error}
+              </div>
             )}
 
-            {/* Success message */}
-            {success && (
-              <motion.div
-                initial={{ opacity: 0, y: -10 }}
-                animate={{ opacity: 1, y: 0 }}
-                className="flex items-start gap-2 p-3 bg-green-50 border border-green-200 rounded-lg text-green-700 text-sm"
-              >
-                <CheckCircle2 className="w-5 h-5 flex-shrink-0 mt-0.5" />
-                <span>{success}</span>
-              </motion.div>
-            )}
-
-            {/* Submit button */}
             <Button
               type="submit"
               fullWidth
               size="lg"
               loading={loading}
-              icon={<ArrowRight className="w-5 h-5" />}
-              iconPosition="right"
             >
-              Sign In
+              Log In
             </Button>
           </form>
 
-          {/* Divider */}
-          <div className="relative my-6">
-            <div className="absolute inset-0 flex items-center">
-              <div className="w-full border-t border-gray-300"></div>
-            </div>
-            <div className="relative flex justify-center text-sm">
-              <span className="px-2 bg-white text-gray-500">Or</span>
-            </div>
-          </div>
-
-          {/* Sign up link */}
-          <div className="text-center">
-            <p className="text-sm text-gray-600">
-              Don't have an account?{' '}
-              <Link href="/signup" className="text-blue-600 hover:underline font-medium">
-                Start your free trial
-              </Link>
-            </p>
-          </div>
+          <p className="mt-6 text-center text-sm text-gray-600">
+            Don't have an account?{' '}
+            <Link href="/signup" className="font-medium text-blue-600 hover:underline">
+              Sign up
+            </Link>
+          </p>
         </div>
-
-        {/* Footer links */}
-        <div className="mt-8 flex items-center justify-center gap-6 text-sm text-gray-500">
-          <Link href="#" className="hover:text-gray-700 transition-colors">
-            Privacy Policy
-          </Link>
-          <span>•</span>
-          <Link href="#" className="hover:text-gray-700 transition-colors">
-            Terms of Service
-          </Link>
-          <span>•</span>
-          <Link href="#" className="hover:text-gray-700 transition-colors">
-            Contact Support
-          </Link>
-        </div>
-      </motion.div>
+      </div>
     </div>
   )
 }
