+++ src/components/SplashScreen.tsx
@@ -2,7 +2,7 @@
 
 import React, { useEffect, useState } from 'react'
 import { motion, AnimatePresence } from 'framer-motion'
-import { Building2, Sparkles } from 'lucide-react'
+import { Building2 } from 'lucide-react'
 
 interface SplashScreenProps {
   onComplete?: () => void
@@ -15,7 +15,7 @@ export function SplashScreen({ onComplete }: SplashScreenProps) {
     const timer = setTimeout(() => {
       setIsLoading(false)
       onComplete?.()
-    }, 2500)
+    }, 2000)
     
     return () => clearTimeout(timer)
   }, [onComplete])
@@ -24,49 +24,15 @@ export function SplashScreen({ onComplete }: SplashScreenProps) {
     <AnimatePresence>
       {isLoading && (
         <motion.div
-          className="fixed inset-0 z-50 flex items-center justify-center bg-gradient-to-br from-blue-50 via-white to-indigo-50"
+          className="fixed inset-0 z-50 flex items-center justify-center bg-white"
           initial={{ opacity: 1 }}
           exit={{ opacity: 0 }}
           transition={{ duration: 0.5 }}
         >
-          {/* Background pattern */}
-          <div className="absolute inset-0 bg-grid-pattern opacity-5" />
-          
-          {/* Animated circles */}
-          <motion.div
-            className="absolute w-96 h-96 bg-blue-400 rounded-full opacity-10 blur-3xl"
-            animate={{
-              scale: [1, 1.2, 1],
-              x: [0, 50, 0],
-              y: [0, -30, 0],
-            }}
-            transition={{
-              duration: 4,
-              repeat: Infinity,
-              ease: "easeInOut"
-            }}
-          />
-          <motion.div
-            className="absolute w-96 h-96 bg-indigo-400 rounded-full opacity-10 blur-3xl"
-            animate={{
-              scale: [1.2, 1, 1.2],
-              x: [0, -50, 0],
-              y: [0, 30, 0],
-            }}
-            transition={{
-              duration: 4,
-              repeat: Infinity,
-              ease: "easeInOut",
-              delay: 0.5
-            }}
-          />
-          
-          {/* Logo and content */}
-          <div className="relative z-10 text-center">
+          <div className="relative flex flex-col items-center justify-center text-center">
             <motion.div
-              className="mb-8"
-              initial={{ scale: 0, rotate: -180 }}
-              animate={{ scale: 1, rotate: 0 }}
+              initial={{ scale: 0.5, opacity: 0 }}
+              animate={{ scale: 1, opacity: 1 }}
               transition={{
                 type: "spring",
                 stiffness: 260,
@@ -74,35 +40,13 @@ export function SplashScreen({ onComplete }: SplashScreenProps) {
                 duration: 0.8
               }}
             >
-              <div className="relative inline-flex items-center justify-center">
-                <motion.div
-                  className="absolute inset-0 bg-gradient-to-r from-blue-600 to-indigo-600 rounded-2xl blur-xl"
-                  animate={{
-                    scale: [1, 1.1, 1],
-                  }}
-                  transition={{
-                    duration: 2,
-                    repeat: Infinity,
-                    ease: "easeInOut"
-                  }}
-                />
-                <div className="relative bg-gradient-to-r from-blue-600 to-indigo-600 p-5 rounded-2xl shadow-2xl">
-                  <Building2 className="w-16 h-16 text-white" />
-                  <motion.div
-                    className="absolute -top-2 -right-2"
-                    initial={{ scale: 0, rotate: -45 }}
-                    animate={{ scale: 1, rotate: 0 }}
-                    transition={{ delay: 0.5, type: "spring" }}
-                  >
-                    <Sparkles className="w-6 h-6 text-yellow-400" />
-                  </motion.div>
-                </div>
+              <div className="bg-blue-600 p-4 rounded-xl shadow-lg">
+                <Building2 className="w-12 h-12 text-white" />
               </div>
             </motion.div>
             
-            {/* Brand name */}
             <motion.h1
-              className="text-4xl md:text-5xl font-bold bg-gradient-to-r from-gray-900 to-gray-700 bg-clip-text text-transparent mb-3"
+              className="text-3xl font-bold text-gray-800 mt-4"
               initial={{ opacity: 0, y: 20 }}
               animate={{ opacity: 1, y: 0 }}
               transition={{ delay: 0.3, duration: 0.5 }}
@@ -110,54 +54,26 @@ export function SplashScreen({ onComplete }: SplashScreenProps) {
               BuildLedger
             </motion.h1>
             
-            {/* Tagline */}
             <motion.p
-              className="text-lg text-gray-600 mb-8"
+              className="text-md text-gray-500"
               initial={{ opacity: 0, y: 20 }}
               animate={{ opacity: 1, y: 0 }}
               transition={{ delay: 0.5, duration: 0.5 }}
             >
-              Financial Command Center for Contractors
+              Quotes, Invoices, Clients. Simplified.
             </motion.p>
             
-            {/* Loading indicator */}
             <motion.div
-              className="flex items-center justify-center gap-2"
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              transition={{ delay: 0.7 }}
-            >
-              <div className="flex gap-1">
-                {[0, 1, 2].map((i) => (
-                  <motion.div
-                    key={i}
-                    className="w-3 h-3 bg-blue-600 rounded-full"
-                    animate={{
-                      y: [0, -10, 0],
-                    }}
-                    transition={{
-                      duration: 0.6,
-                      repeat: Infinity,
-                      ease: "easeInOut",
-                      delay: i * 0.1
-                    }}
-                  />
-                ))}
-              </div>
-            </motion.div>
-            
-            {/* Progress bar */}
-            <motion.div
-              className="mt-8 w-48 h-1 bg-gray-200 rounded-full overflow-hidden mx-auto"
+              className="mt-8 w-32 h-1 bg-gray-200 rounded-full overflow-hidden"
               initial={{ opacity: 0 }}
               animate={{ opacity: 1 }}
               transition={{ delay: 0.8 }}
             >
               <motion.div
-                className="h-full bg-gradient-to-r from-blue-600 to-indigo-600"
+                className="h-full bg-blue-600"
                 initial={{ width: "0%" }}
                 animate={{ width: "100%" }}
-                transition={{ duration: 1.5, ease: "easeInOut", delay: 0.8 }}
+                transition={{ duration: 1.2, ease: "easeInOut", delay: 0.8 }}
               />
             </motion.div>
           </div>
@@ -165,4 +81,4 @@ export function SplashScreen({ onComplete }: SplashScreenProps) {
       )}
     </AnimatePresence>
   )
-} 
+}
